gateway/requirements.txt
# FastAPI & Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.0.0

# Pydantic AI (core of the intelligence layer)
pydantic-ai>=0.1.0
pydantic-ai-slim[logfire]>=0.1.0

# Observability
logfire>=0.50.0

# Temporal for durable workflows
temporalio>=1.7.0

# Voice & AI Services
hume>=0.7.0
zep-cloud>=2.0.0
google-generativeai>=0.8.3

# Database
psycopg[binary]>=3.1.0

# HTTP Client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=24.1.0

# Memory
supermemory>=0.1.0
